requirements/app.txt
guniflask>=0.7.2,<0.8
guniflask-cli>=0.2,<0.3
Flask-SQLAlchemy>=2.4.1
PyMySQL>=0.9.3
